pizza-bot/requirements.txt
python-dotenv
black
ruff
pytest
pg8000
aiohttp
asyncpg
pytest-asyncio